refactor: abstract the SecretExtractor logic to utils
