+++ src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod annotations;
 pub mod line;
+pub mod secret_extractors;
